test(agents): share deletion fixtures
