+++ src/commands/agents.delete.test.ts
@@ -1,30 +1,16 @@
-// Agents delete tests cover config removal, workspace-state cleanup, and binding updates.
 import fs from "node:fs/promises";
 import path from "node:path";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import {
-  listAgentEntries,
-  toAgentEntriesRecord,
-  tryResolveSoleAgentId,
-} from "../agents/agent-scope-config.js";
-import {
   readPersistedAuthProfileStoreRaw,
   writePersistedAuthProfileStoreRaw,
 } from "../agents/auth-profiles/sqlite.js";
-import {
-  retainLegacyDefaultAgentId,
-  tryGetLegacyDefaultAgentId,
-} from "../config/legacy.default-agent-owner.js";
+import { retainLegacyDefaultAgentId } from "../config/legacy.default-agent-owner.js";
 import { resolveSessionStorePathCore } from "../config/sessions.js";
-import type { SessionEntry } from "../config/sessions.js";
-import {
-  listSessionEntriesReadOnly,
-  replaceSessionEntry,
-} from "../config/sessions/session-accessor.js";
+import { listSessionEntriesReadOnly } from "../config/sessions/session-accessor.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import { makeCronJob } from "../cron/delivery.test-helpers.js";
 import { loadCronStore, resolveCronJobsStorePath, saveCronStore } from "../cron/store.js";
-import { GatewayTransportError } from "../gateway/transport-error.js";
 import { readExecApprovalsSnapshot, saveExecApprovals } from "../infra/exec-approvals.js";
 import { parseAgentSessionKey } from "../routing/session-key.js";
 import { readAgentDeletionJournal } from "../state/agent-deletion-journal.js";
@@ -116,6 +102,11 @@ vi.mock("../wizard/clack-prompter.js", () => ({
 }));
 
 import { agentsDeleteCommand } from "./agents.commands.delete.js";
+import {
+  createAgentsDeleteFixture,
+  gatewayTransportError,
+  readAgentDeleteJsonLogs,
+} from "./agents.delete.test-helpers.js";
 
 const runtime = createTestRuntime();
 const sharedAuthStore = {
@@ -125,71 +116,10 @@ const sharedAuthStore = {
   },
 };
 
-function gatewayTransportError(kind: "closed" | "timeout", code?: number): GatewayTransportError {
-  return new GatewayTransportError({
-    kind,
-    code,
-    message: `gateway ${kind}`,
-    connectionDetails: { url: "ws://127.0.0.1:1", urlSource: "test", message: "test gateway" },
-  });
-}
-
-function resolveFixtureStoreAgentId(cfg: OpenClawConfig, deletedAgentId: string): string {
-  const storeConfig = cfg.session?.store;
-  if (typeof storeConfig === "string" && !storeConfig.includes("{agentId}")) {
-    return (
-      tryGetLegacyDefaultAgentId(cfg) ??
-      listAgentEntries(cfg).find((entry) => entry.default === true)?.id ??
-      tryResolveSoleAgentId(cfg) ??
-      deletedAgentId
-    );
-  }
-  return deletedAgentId;
-}
-
-async function arrangeAgentsDeleteTest(params: {
-  stateDir: string;
-  cfg: OpenClawConfig;
-  deletedAgentId?: string;
-  sessions: Record<string, { sessionId: string; updatedAt: number }>;
-}) {
-  const deletedAgentId = params.deletedAgentId ?? "ops";
-  const authored = structuredClone(params.cfg);
-  const roster = listAgentEntries(authored);
-  if (!roster.some((entry) => entry.default === true)) {
-    const existingDefault = roster.find((entry) => entry.id !== deletedAgentId);
-    if (existingDefault) {
-      existingDefault.default = true;
-    } else {
-      roster.unshift({ id: "main", default: true });
-    }
-  }
-  const { list: _legacyList, ...agents } = authored.agents ?? {};
-  const cfg: OpenClawConfig = {
-    ...authored,
-    agents: { ...agents, entries: toAgentEntriesRecord(roster) },
-  };
-  const storeAgentId = resolveFixtureStoreAgentId(cfg, deletedAgentId);
-  const storePath = resolveSessionStorePathCore(cfg.session?.store, { agentId: deletedAgentId });
-  for (const [sessionKey, entry] of Object.entries(params.sessions)) {
-    const entryAgentId = parseAgentSessionKey(sessionKey)?.agentId ?? storeAgentId;
-    const entryStorePath = resolveSessionStorePathCore(cfg.session?.store, {
-      agentId: entryAgentId,
-    });
-    await replaceSessionEntry({ agentId: entryAgentId, sessionKey, storePath: entryStorePath }, {
-      ...entry,
-      delivery: { kind: "none" },
-    } as SessionEntry);
-  }
-  await fs.mkdir(path.join(params.stateDir, `workspace-${deletedAgentId}`), { recursive: true });
-  await fs.mkdir(path.join(params.stateDir, "agents", deletedAgentId, "agent"), {
-    recursive: true,
-  });
-
+const arrangeAgentsDeleteTest = createAgentsDeleteFixture((cfg) => {
   configMocks.readConfigFileSnapshot.mockResolvedValue(createTestConfigSnapshot(cfg));
-
-  return storePath;
-}
+});
+const readJsonLogs = () => readAgentDeleteJsonLogs(runtime.log.mock.calls);
 
 function expectSessionStore(
   cfg: OpenClawConfig,
@@ -220,15 +150,6 @@ function expectSessionStore(
       ]),
     ),
   );
-}
-
-function readJsonLogs(): Array<Record<string, unknown>> {
-  return runtime.log.mock.calls
-    .filter((call): call is [string, ...unknown[]] => {
-      const arg = call[0];
-      return typeof arg === "string" && arg.startsWith("{");
-    })
-    .map((call) => JSON.parse(call[0]) as Record<string, unknown>);
 }
 
 describe("agents delete command", () => {
